[feat] Improve the friendly names for PAs.
This commit moves the PA friendly name definitions into
_config.yml, and updates breadcrumbs and the summary boxes
to use them.

This also gets rid of the parse-map-from-string-on-the-fly
thing that was caused by lack of useful collection support
in Liquid. Since we are using this in multiple files now,
it makes sense for it to go to `_config.yml`.

Also changed the `client-libraries` scope to kebob-case
since we are no longer using the scope name for display.