[AIP-0133] Update guidelines for user-specified IDs
Ignoring resource.name on create can be a confusing user experience.
This change proposes a modification to the guidelines where the name
is validated as empty or must match the name that the service creates.

An alternative is to also allow resource_id to be empty when users
supply a valid name. This makes it easier for consumers of the API
to get it right, but has the downside of providing two correct ways
of doing something.