Change examples from "shelves" to "publishers".
There is broad agreement that `shelves/*/books/*` is a poor
example of a resource hierarchy, as moving a book from one shelf
to another is a common operation and so a shelf should be a
property of a Book resource, not a parent. In fact, we often
give disclaimers about this when teaching the API design class.

Publishers is a better natural parent for a book. A book can
(and does) theoretically change publishers, but it is an expensive
operation and it also makes sense that its resource name would
change in this situation (the new book would have a new ISBN,
etc. -- it is really an equivalent but not identical book once
this happens).