[feat] Remove the state column, replace with a badge.
This PR removes the "state" column from the
general AIP index and replaces it with a
"Reviewing" or "Draft" badge for AIPs in those
states (and removes other states from the list
entirely).

This change only affects the "general" page to
avoid needing everyone from every team to approve.
After this is merged, I will send individual PRs
to each team to migrate their indexes.