[fix] Remove /home from the user click flow.
Landing on aip.dev, you see a big blue button that says "Explore
the AIPs", but it actually takes you to a page that largely
regurgitates the same information as the home page, with what you
expect -- the list of AIPs -- on the left sidebar.

This short-circuits that: the "explore" button goes directly to
the general AIP listing.