[docs] AIP newsletter - March 2020 (#423)
* [docs] AIP newsletter - March 2020

This is my almost-complete stab at offering a monthly newsletter.
The idea would be that we send this internally near the beginning
of each month to invite comment on pending proposals and share
other information. It will also serve as an overall changelog for
important changes (new AIPs, not minor changes to existing ones).

The way I plan to do this is to make a milestone for each month
containing both the newsletter itself and any applicable PRs that
may need comment.