[refactor] Stop forking all of Glue.
This PR removes the fork of the Glue SCSS. We were previously
forking the entire file in order to be able to use `@extend`
directives, but this is adding significant overhead. It also
creates perverse CSS where every Glue directive is repeated,
which requires more browser processing.

Instead, we backdoor the extend directives using JavaScript
and add the classes we need explicitly there.

- This reduces the build time from 40 seconds to 14 seconds.
- The change in render speed is not discernible. I expect it
  may actually be faster because the 7-8 additional jQuery
  commands are offset by a 40% reduction in total CSS to process.