[fix] Fix search pagination. (#484)
Uses #483 as a base.

I noticed when working on #483 that the search pagination was
broken, probably from going from Glue 19 to Glue 22. Glue's
pagination is now much more than CSS and is untenable (because we
do not control what HTML Tipue uses and Tipue rewrites it
constantly), so this sets up nice looking pagination that works
with what we have.