AIP-162: Forbid deleting only remaining revision
The guidance for AIP-162 indicates that a revisioned resource should at all times have at least one revision, since it should always have a non-empty value with which to populate the revision_id field.

The behavior of Delete is to destroy the resource entirely, including any child resources if the optional 'force' flag is set. DeleteRevision doesn't indicate that the 'force' flag is permitted, which suggests that DeleteRevision can't delete a resource entirely.

Allowing DeleteRevision to delete all revisions of a resource could circumvent a policy preventing the user from deleting the resource directly.