AIP-211: Add a Rationale section.
The cross-company AIP team debated this AIP at length, and decided
to endorse the current Google position more widely. During this
debate, the group fleshed out a much more detailed rationale, which
we believe should live within the AIP in this case.

This PR ports the cross-company rationale section (currently in a
private repo) back to the Google AIP.