Clarify scope of common protos (#836)
* Use googleapis repo for common protos instead of api-common-protos

There is a long-standing redundancy between api-common-protos and googleapis. It looks like we're coming down on the side of archiving api-common-protos (which is already out of date) so let's make sure the AIPs reference the "live" versions.

Note that this *doesn't* change the container name in AIP-4290. I'm consulting on that separately.

* Clarify scope of common protos