Fix the descriptions of suffix to be more precise (#903)
* Make the description of suffix use more precise

The existing suffix descriptions of "-Request" or "-Response" aren't quite accurate: the required suffix is simply "Request" or "Response". Including the "-" in the description is superfluous, as it is already clearly described as a "suffix", and potentially confusing as the reader could understand the "-Request" or "-Response" to be verbatim and inconsistent with the examples. (Also note that later in this doc the ":" symbol is given to be verbatim.)

* add changelog item

* add updated field to header

* remove "-" from all Request and Response suffixes