Make CRUDL rpcs call resource field "resource"
In the current CRUDL methods, it's difficult to determine which field in the
request refers to the resource.

e.g. for the example: a code generator would have to do the following
to determine the field containing the resource is "book":

1. read the message
2. strip the operation prefix, which has to be inferred from the parent
  operation referring to the resource.
3. strip the suffix "Request/Response"

Compare this to a static string that refers to the same concept.

This change will require change in the AIP linter:

- https://linter.aip.dev/134/request-resource-field
- https://linter.aip.dev/133/request-resource-field