move guidance for resource ref name out of note
The guidance around resource reference using the resource name
to identify the resource was in a note below the proto example,
which can be easily missed.

Moving the guidance specific to ID to it's own line item, and
clarifying in the section that resource references should
use the resource name to refer to another resource.