feat(AIP-122) add guidance to prefer easy plurals
When referring to an API collection, it's common to need to
refer to the collection in both the plural and singular.

Plural and singular and English can be inconsistent in it's
conversion: some words have different endings (butterflies
vs butterfly), while others are the same word (moose).

If most terms did have a straightforward conversion across
singular and plural, it would reduce the need for a user's
awareness of the English language, as well as enable some
heuristic that could be applied for automation.