fix: use the term uniformity instead of consistency
The term "consistency" is overloaded, mixing the compliance of
multiple things to a particular standard or convention, vs the
computer science term with respect to data state across replicas.

Using the term "uniformity" conveys the same intention, and provides
a clear delineation between the terms.

In some places "consistency" is a better term than uniformity, so not
all instances were replaced.

Similarly, not replacing the term from any news articles as those
are point-in-time artifacts.