feat(121) clarifying one std method per resource
Although implied, it was not stated that there can only
be one standard method per resource.

Having multiple standard methods for a single resources can make
it impossible for automated clients to reason about the appropriate
call to operate on a resource. This impacts declarative clients
heavily, which need to determine which mutation method is appropriate.