feat(AIP-133) up resource_id guidance to should
The current wording implies that user-settable resource_ids
are somewhat helpful *may*, when in most cases they are very
beneficial: they are critical for declarative clients to uniquely
identify resources and prevent duplicates, and enable the user
to provide semantic names and organization for resources.

Generally the guidance would be better for APIs if they were
added if at all possible, and only skipped if there the usage
pattern is exceptional (e.g. job resources with significant churn,
revisions where a unique identifier is determined by the custom
method).