Update 0126.md
The current AIP around naming and prefixing enums contains some confusing instructions. In particular, we currently say that "other values **should not** be prefixed"  but also "To avoid sharing values, APIs **may** prefix". The suggestion here aims to clarify when a value should be prefixed and when it shouldn't. We want to clarify that:
1. Nested enums should not be prefixed.
2. Package level enums should be prefixed.