feat(AIP-130): adding method AIP
There's a lot of confusion around the types of methods that are
possible to service teams, as well as which ones they should be using.

This is the start of an AIP to try to codify that a bit better.

The AIP starts in draft state, as I think it needs more vetting and
time to determine if the terminology is correct. At the same time,
there is value in adding them to the main AIPs so that they are
more visible and we can contribute incremental improvements.

This is not comprehensive, and is meant to reach some agreement
on the general structure of the AIP.