feat(AIP-128): remove unused guidance (#1009)
There is guidance in AIP-128 that is currently not
relied on by any declarative surface.

This is not a comprehensive removal of all unused behavior,
but it does remove guidance this generally unhelpful.

Specifically:

Delete if existing is not necessary: error codes exist to indicate
if a resource has already been deleted, and can be used by declarative
surfaces instead.

For soft deletes: this must only be used in the case where the
resource cannot re-use the same id. Generally it is a simpler
workflow to simple recreate the resource rather than undelete.

* addressing feedback

switching soft-delete guidance to *should not*, explaining the only
caveats.