feat(AIP-111): introducing plane AIP
The term data plane have been used in the AIPs without a clear
definition.

Introducing the term "management" and "data" plane, as these are
industry-standard terms and are also well defined in the context of
Google APIs.

These definitions are also helpful in:

- providing fine-grained API guidance depending on the plane
- help inform new annotations on resources and methods, which can be
  used for auditing and determining coverage
- further help clarify classifications of methods (AIP-130)

Also removing the redundant text in AIP-128.