feat(AIP-203): introduce IDENTIFIER field behavior
Identifier fields are often multi-modal: they are required or optional
in various situations. Adding an explicit field behavior for these
fields helps eliminate ambiguity for an author to fully annotate field
behavior on all request fields, as well as clarify which fields to use
in identifying the resource for clients.

fixes #1176.