fix: links from "general" should be relative in effective output directory
Even though in the source there's a "general" directory (so ../client-libraries/ makes sense) that isn't the case when built... which I assume is why none of these links work.
I'm not sure the best way to test this - if it doesn't work, we may need to just use absolute URLs, like the ones to the auth AIPs.

Fixes #1394