AIP-121: List responses contain the resource
AIP-132 documents the guidance for List responses, which *contain* the
resource as a `repeated` field, along with pagination information. This
patch updates AIP-121 to consistently indicate that the List request
merely contains the resource rather than being the resource, which
wouldn't make sense since the point of a List request is to enumerate
multiple such resources.