feat(AIP-141): add naming guidance compound and/or inverse unit fields (#1524)
* Updates AIP-141 to detail guidance for compound and inverse units

This includes a new subsection covering compound units, a new subsection covering inverse units, and some overall wording adjustments to clarify when plural form should be used for field names.

* Update AIP-140 to reference AIP-141 for use of preposition "per"

AIP-141 has been updated with more detailed guidance on the use of the preposition "per" in field names representing quantities, and this note should be updated to refer to that guidance.

Note that the previous note guidance here may have suggested that an event-frequency field should be named like e.g. "failures_per_hour", while after this change it will point to guidance in AIP-141 that specifies the format as e.g. "failure_count_per_hour". This may be interpreted as a functional change.