Update guidance on URL/URI naming in AIP-140 ✍️
This proposes to update the guidelines to advocate for clearer naming for fields that can only return URLs.

It appears to be possible to misinterpret AIP-0140 guidelines to infer that fields that return URLs MUST be named as if they could be URIs. It seems more accurate to indicate that fields that COULD be URIs should use `uri` in the name and fields that MUST be URLs should use `url` in the name.